+++ aloe/utils.py
@@ -56,13 +56,10 @@ else:
             and unicode.
             """
 
-            # pylint:disable=super-on-old-class
-            # https://bitbucket.org/logilab/pylint/issues/721
             try:
                 super().write(str_)
             except TypeError:
                 super().write(str_.decode('utf-8'))
-            # pylint:enable=super-on-old-class
 
 
 def unwrap_function(func):